+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed the `parse_descriptor` example, since it wasn't demostrating any bdk-specific API anymore.
 #### Changed
 - Updated `bitcoin` to `0.26`, `miniscript` to `5.1` and `electrum-client` to `0.6`
+#### Added
+- Added support for the `signet` network (issue #62)
 
 #### Added
 - Added a function to get the version of BDK at runtime